requirements.txt
streamlit
google-generativeai
python-dotenv
langchain_google_genai
gunicorn
langchain_community
# openai
google-cloud-aiplatform>=1.38
# ---------------------------------------
#sklearn
#numpy==1.24.3
numpy
plotly
matplotlib
#matplotlib==3.3.4
mysql-connector-python
bcrypt
seaborn
pillow
# ---------------------------------------

PyPDF2
faiss-cpu
groq
langchain-groq
pypdf

# --------------------------------

python-magic
python-docx

# --------------------------------
transformers
torch